big refactor: saved old logic in backup folders and removed disabled pages
